ComputerVision/TransferLearning/code/requirements.txt
tqdm
pytorch-lightning
sagemaker
boto3
torchvision
torchsummary
tqdm
torchmetrics